+++ contracts/market/Market.sol
@@ -33,7 +33,7 @@ contract Market is IMarket, CommonContext, MarketSettingsContext, AccessControlE
     using SafeCast for int;
 
     // reserved storage slots for base contract upgrade in future
-    uint256[50] private __gap;
+    uint[50] private __gap;
 
     // states
     address public baseToken; // liquidity token
