+++ contracts/oracle/PriceOracle.sol
@@ -24,7 +24,7 @@ contract PriceOracle is IPriceOracle, MarketSettingsContext, AccessControlEnumer
     uint public constant PRICE_PRECISION = 18;
 
     // reserved storage slots for base contract upgrade in future
-    uint256[50] private __gap;
+    uint[50] private __gap;
 
     //states
     // chainlink price feed aggregators
