+++ contracts/tokens/SYM.sol
@@ -29,7 +29,7 @@ contract SYM is ISYM, ERC20, ERC20Pausable, AccessControlEnumerable, PauseContro
     function _beforeTokenTransfer(
         address from,
         address to,
-        uint256 amount
+        uint amount
     ) internal virtual override(ERC20, ERC20Pausable) {
         super._beforeTokenTransfer(from, to, amount);
     }
